+++ CHANGELOG.md
@@ -4,6 +4,8 @@
     instead of `current_user` (if `current_user` is defined).
   - [#313](https://github.com/airblade/paper_trail/pull/313) - Make the `Rails::Controller` helper compatible with
     `ActionController::API` for compatibility with the [`rails-api`](https://github.com/rails-api/rails-api) gem.
+  - Deprecated `Model.paper_trail_on` and `Model.paper_trail_off` in favor of bang versions of the methods. Deprecation warning
+    informs users that the non-bang versions of the methods will be removed in version `3.1.0`.
 
 ## 3.0.0